+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/PlayerMixin.java
@@ -1,14 +1,17 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.iguanatweaksreborn.event.ITREventFactory;
 import insane96mcp.iguanatweaksreborn.module.combat.stats.Stats;
 import insane96mcp.iguanatweaksreborn.module.experience.PlayerExperience;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
@@ -18,6 +21,7 @@ import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import javax.annotation.Nullable;
+import java.util.Map;
 
 //Higher priority over ShieldsPlus. This makes this run first so ShieldPlus overrides this.
 @Mixin(Player.class)
@@ -80,5 +84,15 @@ public abstract class PlayerMixin extends LivingEntity {
 	@ModifyConstant(method = "attack", constant = @Constant(floatValue = 0.8f, ordinal = 0))
 	public float attackStrengthAtFullSwing(float value) {
 		return Stats.noDamageWhenSpamming() ? 1f : value;
+	}
+
+
+	@ModifyExpressionValue(method = "attack", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/enchantment/EnchantmentHelper;getDamageBonus(Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/entity/MobType;)F"))
+	public float onEnchantmentDamage(float original, Entity target) {
+		Map<Enchantment, Integer> allEnchantments = this.getMainHandItem().getAllEnchantments();
+		for (Enchantment enchantment : allEnchantments.keySet()) {
+			original += EnchantmentsFeature.bonusDamageEnchantment(enchantment, allEnchantments.get(enchantment), this, target);
+		}
+		return original;
 	}
 }