+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/damage/Smite.java
@@ -4,9 +4,9 @@ import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.LivingEntity;
 
 public class Smite extends BonusDamageEnchantment {
     public static final TagKey<EntityType<?>> AFFECTED_BY_SMITE = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(IguanaTweaksReborn.MOD_ID, "enchanting/smite"));
@@ -15,7 +15,7 @@ public class Smite extends BonusDamageEnchantment {
     }
 
     @Override
-    public boolean isAffectedByEnchantment(LivingEntity target) {
+    public boolean isAffectedByEnchantment(Entity target) {
         return target.getType().is(AFFECTED_BY_SMITE);
     }