+++ changelog.md
@@ -7,6 +7,7 @@
   * Increased spawn speed (every 25\~100 -> 20\~80 seconds)
 * Increased Hoes base attack speed bonus (+15% -> +30%)
 * Possibly fixed chunks staying loaded after player second death
+* Fixed damaging enchantments (e.g. Smite) applying to sweep attack
 
 ## 3.8.4
 * Stamina now unlocks at 50% instead of 75%